+++ src/mruby_integration/models/image.cpp
@@ -435,6 +435,26 @@ mrb_Image_greyscale_bang(mrb_state* mrb, mrb_value self) -> mrb_value
 }
 
 auto
+mrb_Image_contrast_bang(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  Image* image;
+
+  Data_Get_Struct(mrb, self, &Image_type, image);
+  mrb_assert(image != nullptr);
+
+  mrb_float contrast;
+  mrb_get_args(mrb, "f", &contrast);
+
+  if (contrast < -100 || contrast > 100) {
+    mrb_raise(mrb, E_ARGUMENT_ERROR, "Must be within (-100..100)");
+  }
+
+  ImageColorContrast(image, contrast);
+
+  return self;
+}
+
+auto
 mrb_Image_get_data(mrb_state* mrb, mrb_value self) -> mrb_value
 {
   Image* image;
@@ -550,6 +570,8 @@ append_models_Image(mrb_state* mrb)
     mrb, Image_class, "invert!", mrb_Image_invert_bang, MRB_ARGS_NONE());
   mrb_define_method(
     mrb, Image_class, "greyscale!", mrb_Image_greyscale_bang, MRB_ARGS_NONE());
+  mrb_define_method(
+    mrb, Image_class, "contrast!", mrb_Image_contrast_bang, MRB_ARGS_REQ(1));
   mrb_define_method(
     mrb, Image_class, "data", mrb_Image_get_data, MRB_ARGS_NONE());
   mrb_define_method(